requirements.txt
PyGithub
gql
PyYAML